refactor: use anyhow for errors
This employs `anyhow::Error` for all errors.
This is due to the use of `FromStr` instead of a
separate trait, which simplifies the API, but
allows anyone to implement parse errors, which
has to be properly handled.

BREAKING CHANGE: the previously used error
was removed in favor or `anyhow::Error`.